fix: some minor fix
